backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langgraph
langsmith
groq
deepgram-sdk
httpx
pydantic
python-multipart
